+++ core/dto/place.dto.go
@@ -7,8 +7,8 @@ type PlaceDto struct {
 	PaymentNeed bool         `json:"payment_need"`
 	Title       string       `json:"title"`
 	Description string       `json:"description"`
-	Longitude   float32      `json:"longitude"`
-	Latitude    float32      `json:"latitude"`
+	Longitude   float64      `json:"longitude"`
+	Latitude    float64      `json:"latitude"`
 	CreatedBy   BaseUserInfo `json:"created_by"`
 	RegionInfo  RegionDto    `json:"region_info"`
 	PlacePhotos []string     `json:"photos"`
@@ -18,8 +18,8 @@ type CreatePlace struct {
 	PaymentNeed bool    `json:"payment_need"`
 	Title       string  `json:"title" binding:"required"`
 	Description string  `json:"description" binding:"required"`
-	Longitude   float32 `json:"longitude" binding:"required"`
-	Latitude    float32 `json:"latitude" binding:"required"`
+	Longitude   float64 `json:"longitude" binding:"required"`
+	Latitude    float64 `json:"latitude" binding:"required"`
 	RegionID    string  `json:"region_id" binding:"required"`
 }
 
@@ -27,7 +27,7 @@ type UpdatePlace struct {
 	PaymentNeed bool    `json:"payment_need"`
 	Title       string  `json:"title"`
 	Description string  `json:"description"`
-	Longitude   float32 `json:"longitude"`
-	Latitude    float32 `json:"latitude"`
+	Longitude   float64 `json:"longitude"`
+	Latitude    float64 `json:"latitude"`
 	RegionID    string  `json:"region_id"`
 }
