+++ core/repo/place.repo.go
@@ -14,8 +14,8 @@ type Place struct {
 	CreatedByID     uuid.UUID `gorm:"not null"`
 	Title           string    `gorm:"not null"`
 	Description     string    `gorm:"not null"`
-	Longitude       float32   `gorm:"not null"`
-	Latitude        float32   `gorm:"not null"`
+	Longitude       float64   `gorm:"not null"`
+	Latitude        float64   `gorm:"not null"`
 	RegionID        uuid.UUID
 
 	Region    Region `gorm:"foreignKey:RegionID;references:ID"`
